Added natsort, analysis of numpy array to prep for checking markers
Natural sort added to overcome issues with natural sorting algorithm
While working on marker checks (assuring they correspond to actual cells
in the stack), discovered that RGB values held at the location in the numpy
array don't correspond to the same values in the stack. Not sure what
the issue could be, and this is holding me up quite a bit.